Refactor timing mechanism by injecting Stopwatch
The timing code has been refactored to make it more precise.
Previously, the timing included the function call as well as any
checks to determine if the method was successful (issets() and array
checks mostly).  This is not technically accurate, as that is
test validation, not the library itself.

All clients have been modified to accept the StopWatch object
for each method, so that they can call start/stop themselves.  This
allows much finer control over the timing.  The object is passed
by-reference.

Similarly, the timing of the entire client was very inaccurate,
as it included instantiation time of both version of clients
in the constructor, as well as any error handling and bounds checking.

More accurate is to just sum up the individual method times.

Also added some prettier formatting