Add Persistent vs Transient tests
This creates a new test type called "persistent".  The difference
is how the client is instantiated.  In the old ("transient") tests,
the client was rebuilt on every request.  This is a great way to
profile use-cases where you need to instantiate a client
quickly and fire off a few queries (e.g. autocomplete, suggestions,
even regular search-boxes).

On the other hand, this doesn't accurately represent cases where
a client is required to maximize throughput over a large number
of queries, such as bulk importing documents or performing
heavy analytics.

The new "persistent" test instantiates a single client and uses
that for all tests.