Wait for yellow cluster state
Due to the automated speed of index creation + document insertion,
it is possible for an index to be not fully initialized before
the script tries to add documents.

This adds an explicit "wait for yellow" call so that the index
is fully initialized before continuing.

Also removes replicas, since they are meaningless on a one-node
setup.